+++ src/core/utils/templating/utils.ts
